requirements.txt
requests
django
pytest
pytest-django
djangorestframework
djangorestframework-simplejwt
flake8
isort
django-filter
